requirements.txt
deepl==1.14.0
discord.py==2.2.2
cs-board-tools @ git+https://github.com/FortuneStreetModding/cs-board-tools@main
